standardize: to float64

+++ src/models/physics/bmi_cfe.py
@@ -236,59 +236,61 @@ class BMI_CFE:
         # ________________________________________________
         # Inputs
         self.timestep_rainfall_input_m = torch.zeros(
-            (1, self.num_basins), dtype=torch.float
+            (1, self.num_basins), dtype=torch.float64
         )
-        self.potential_et_m_per_s = torch.zeros((1, self.num_basins), dtype=torch.float)
+        self.potential_et_m_per_s = torch.zeros(
+            (1, self.num_basins), dtype=torch.float64
+        )
 
         # ________________________________________________
         # flux variables
         self.flux_overland_m = torch.zeros(
-            (1, self.num_basins), dtype=torch.float
+            (1, self.num_basins), dtype=torch.float64
         )  # surface runoff that goes through the GIUH convolution process
         self.flux_perc_m = torch.zeros(
-            (1, self.num_basins), dtype=torch.float
+            (1, self.num_basins), dtype=torch.float64
         )  # flux from soil to deeper groundwater reservoir
         self.flux_lat_m = torch.zeros(
-            (1, self.num_basins), dtype=torch.float
+            (1, self.num_basins), dtype=torch.float64
         )  # lateral flux in the subsurface to the Nash cascade
         self.flux_from_deep_gw_to_chan_m = torch.zeros(
-            (1, self.num_basins), dtype=torch.float
+            (1, self.num_basins), dtype=torch.float64
         )  # flux from the deep reservoir into the channels
         self.gw_reservoir_storage_deficit_m = torch.zeros(
-            (1, self.num_basins), dtype=torch.float
+            (1, self.num_basins), dtype=torch.float64
         )  # the available space in the conceptual groundwater reservoir
         self.primary_flux = torch.zeros(
-            (1, self.num_basins), dtype=torch.float
+            (1, self.num_basins), dtype=torch.float64
         )  # temporary vars.
         self.secondary_flux = torch.zeros(
-            (1, self.num_basins), dtype=torch.float
+            (1, self.num_basins), dtype=torch.float64
         )  # temporary vars.
         self.primary_flux_from_gw_m = torch.zeros(
-            (1, self.num_basins), dtype=torch.float
+            (1, self.num_basins), dtype=torch.float64
         )
         self.secondary_flux_from_gw_m = torch.zeros(
-            (1, self.num_basins), dtype=torch.float
+            (1, self.num_basins), dtype=torch.float64
         )
-        self.total_discharge = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.diff_infilt = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.diff_perc = torch.zeros((1, self.num_basins), dtype=torch.float)
+        self.total_discharge = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.diff_infilt = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.diff_perc = torch.zeros((1, self.num_basins), dtype=torch.float64)
 
         # ________________________________________________
         # Evapotranspiration
         self.potential_et_m_per_timestep = torch.zeros(
-            (1, self.num_basins), dtype=torch.float
+            (1, self.num_basins), dtype=torch.float64
         )
         self.actual_et_m_per_timestep = torch.zeros(
-            (1, self.num_basins), dtype=torch.float
+            (1, self.num_basins), dtype=torch.float64
         )
         self.reduced_potential_et_m_per_timestep = torch.zeros(
-            (1, self.num_basins), dtype=torch.float
+            (1, self.num_basins), dtype=torch.float64
         )
         self.actual_et_from_rain_m_per_timestep = torch.zeros(
-            (1, self.num_basins), dtype=torch.float
+            (1, self.num_basins), dtype=torch.float64
         )
         self.actual_et_from_soil_m_per_timestep = torch.zeros(
-            (1, self.num_basins), dtype=torch.float
+            (1, self.num_basins), dtype=torch.float64
         )
 
         # ________________________________________________
@@ -341,7 +343,7 @@ class BMI_CFE:
         # lateral flow function parameters
         assumed_near_channel_water_table_slope = 0.01  # [L/L]
         lateral_flow_threshold_storage_m = field_capacity_storage_threshold_m
-        self.soil_reservoir_storage_deficit_m = torch.tensor([0.0], dtype=torch.float)
+        self.soil_reservoir_storage_deficit_m = torch.tensor([0.0], dtype=torch.float64)
 
         # ________________________________________________
         # Subsurface reservoirs
@@ -351,14 +353,14 @@ class BMI_CFE:
             "coeff_primary": self.Cgw,
             "exponent_primary": self.expon,
             "storage_threshold_primary_m": torch.zeros(
-                (1, self.num_basins), dtype=torch.float
+                (1, self.num_basins), dtype=torch.float64
             ),
             # The following parameters don't matter. Currently one storage is default. The secoundary storage is turned off.
             "storage_threshold_secondary_m": torch.zeros(
-                (1, self.num_basins), dtype=torch.float
+                (1, self.num_basins), dtype=torch.float64
             ),
-            "coeff_secondary": torch.zeros((1, self.num_basins), dtype=torch.float),
-            "exponent_secondary": torch.ones((1, self.num_basins), dtype=torch.float),
+            "coeff_secondary": torch.zeros((1, self.num_basins), dtype=torch.float64),
+            "exponent_secondary": torch.ones((1, self.num_basins), dtype=torch.float64),
         }
         self.gw_reservoir["storage_m"] = self.gw_reservoir["storage_max_m"] * 0.01
         self.volstart = self.volstart.add(self.gw_reservoir["storage_m"])
@@ -372,12 +374,12 @@ class BMI_CFE:
             * self.soil_params["slop"]
             * self.time_step_size,  # Controls percolation to GW, Equation 11
             "exponent_primary": torch.ones(
-                (1, self.num_basins), dtype=torch.float
+                (1, self.num_basins), dtype=torch.float64
             ),  # Controls percolation to GW, FIXED to 1 based on Equation 11
             "storage_threshold_primary_m": field_capacity_storage_threshold_m,
             "coeff_secondary": self.K_lf,  # Controls lateral flow
             "exponent_secondary": torch.ones(
-                (1, self.num_basins), dtype=torch.float
+                (1, self.num_basins), dtype=torch.float64
             ),  # Controls lateral flow, FIXED to 1 based on the Fred Ogden's document
             "storage_threshold_secondary_m": lateral_flow_threshold_storage_m,
         }
@@ -391,9 +393,11 @@ class BMI_CFE:
             self.refkdt * self.satdk / 2.0e-06
         )
         self.Schaake_output_runoff_m = torch.zeros(
-            (1, self.num_basins), dtype=torch.float
+            (1, self.num_basins), dtype=torch.float64
         )
-        self.infiltration_depth_m = torch.zeros((1, self.num_basins), dtype=torch.float)
+        self.infiltration_depth_m = torch.zeros(
+            (1, self.num_basins), dtype=torch.float64
+        )
 
         # ________________________________________________________
         self.runoff_queue_m_per_timestep = torch.zeros(
@@ -448,38 +452,44 @@ class BMI_CFE:
     # ________________________________________________
     # Mass balance tracking
     def reset_volume_tracking(self):
-        self.volstart = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_et_from_soil = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_et_from_rain = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_partition_runoff = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_partition_infilt = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_out_giuh = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_end_giuh = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_to_gw = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_to_gw_start = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_to_gw_end = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_from_gw = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_in_nash = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_in_nash_end = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_out_nash = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_soil_start = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_to_soil = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_soil_to_lat_flow = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_soil_to_gw = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_soil_end = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.volin = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.volout = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.volend = torch.zeros((1, self.num_basins), dtype=torch.float)
+        self.volstart = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.vol_et_from_soil = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.vol_et_from_rain = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.vol_partition_runoff = torch.zeros(
+            (1, self.num_basins), dtype=torch.float64
+        )
+        self.vol_partition_infilt = torch.zeros(
+            (1, self.num_basins), dtype=torch.float64
+        )
+        self.vol_out_giuh = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.vol_end_giuh = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.vol_to_gw = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.vol_to_gw_start = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.vol_to_gw_end = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.vol_from_gw = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.vol_in_nash = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.vol_in_nash_end = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.vol_out_nash = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.vol_soil_start = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.vol_to_soil = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.vol_soil_to_lat_flow = torch.zeros(
+            (1, self.num_basins), dtype=torch.float64
+        )
+        self.vol_soil_to_gw = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.vol_soil_end = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.volin = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.volout = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.volend = torch.zeros((1, self.num_basins), dtype=torch.float64)
         self.vol_partition_runoff_IOF = torch.zeros(
-            (1, self.num_basins), dtype=torch.float
+            (1, self.num_basins), dtype=torch.float64
         )
         self.vol_partition_runoff_SOF = torch.zeros(
-            (1, self.num_basins), dtype=torch.float
+            (1, self.num_basins), dtype=torch.float64
         )
-        self.vol_et_to_atm = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_et_from_soil = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_et_from_rain = torch.zeros((1, self.num_basins), dtype=torch.float)
-        self.vol_PET = torch.zeros((1, self.num_basins), dtype=torch.float)
+        self.vol_et_to_atm = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.vol_et_from_soil = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.vol_et_from_rain = torch.zeros((1, self.num_basins), dtype=torch.float64)
+        self.vol_PET = torch.zeros((1, self.num_basins), dtype=torch.float64)
         return
 
     # ________________________________________________________
@@ -489,7 +499,7 @@ class BMI_CFE:
 
         # the GIUH queue might have water in it at the end of the simulation, so sum it up.
         self.vol_end_giuh = torch.sum(
-            self.runoff_queue_m_per_timestep, dtype=torch.float
+            self.runoff_queue_m_per_timestep, dtype=torch.float64
         )
         self.vol_in_nash_end = torch.sum(self.nash_storage)
